FastQ gzip parsing and list of infiles working, dict comprehensions replaced for compat with 2.6

+++ sieve/multirunner.py
@@ -1,22 +1,32 @@
 from __future__ import print_function
 import os
+
 from util.combinations import combinations
+from sieve import Sieve
 
-
-def init(self, params):
-    self.param_names = [
-        'sieve',
-        'params'
-    ]
+class MultiRunner(Sieve):
+    def init(self, params):
+        self.indbmode = True
+        self.outdbmode = True
+        self.name = 'MultiRunner'
+        self.param_names = [
+            'sieve',
+            'params'
+        ]
 
     def run(self, indnadb, inprotdb, infilepath, outdnadb, outprotdb, outfilepath):
         (outdir, outfile) = outfilepath.rsplit('/', 1)
         for params in combinations(self.params):
-            soutdir = '/'.join([outdir,  '_'.join(str(x[1]) for x in params)])
-            if not os.pathexists(soutdir):
+            soutdir = '/'.join([outdir,  '_'.join(str(x) for x in params.values())])
+            if not os.path.exists(soutdir):
                 os.makedirs(soutdir)
-            outdnadb.truncate()
-            outprotdb.truncate()
-            self.sieve(indnadb, inprotdb, infilepath, outdnadb, outprotdb, '/'.join[soutdir, outfile])
 
+            if outdnadb is not None:
+                outdnadb.truncate()
+            if outprotdb is not None:
+                outprotdb.truncate()
+            s = self.sieve.sieve.create(params, self.logfile)
+            soutfile = '/'.join([soutdir, outfile])
+            s.run(indnadb, inprotdb, infilepath, outdnadb, outprotdb, soutfile)
 
+sieve = MultiRunner
